steel_worldgen::density_functions

Module overworld

Overworld density functions.

Structs

OverworldColumnCache

Column-level cache for flat-cached (xz-only) density function results.

OverworldNoiseSettings

Noise settings for this dimension, parsed from the datapack.

OverworldNoises

All noise generators needed by this dimension’s density functions.

Constants

INTERPOLATED_COUNT

Total number of independently interpolated channels across all router entries (final_density + vein_toggle + vein_ridged).

VEIN_INTERP_ENABLED

Whether vein functions have interpolation channels.

Functions

combine_interpolated

Combine interpolated values for final_density.

combine_vein_ridged

Combine interpolated values for vein_ridged.

combine_vein_toggle

Combine interpolated values for vein_toggle.

compute_overworld__base_3d_noise 🔒

minecraft:overworld/base_3d_noise

compute_overworld__base_3d_noise_4x 🔒

minecraft:overworld/base_3d_noise (SIMD form, batches 4 Y values)

compute_overworld__caves__entrances 🔒

minecraft:overworld/caves/entrances

compute_overworld__caves__entrances_4x 🔒

minecraft:overworld/caves/entrances (SIMD form, batches 4 Y values)

compute_overworld__caves__noodle 🔒

minecraft:overworld/caves/noodle

compute_overworld__caves__noodle_4x 🔒

minecraft:overworld/caves/noodle (SIMD form, batches 4 Y values)

compute_overworld__caves__pillars 🔒

minecraft:overworld/caves/pillars

compute_overworld__caves__pillars_4x 🔒

minecraft:overworld/caves/pillars (SIMD form, batches 4 Y values)

compute_overworld__caves__spaghetti_2d 🔒

minecraft:overworld/caves/spaghetti_2d

compute_overworld__caves__spaghetti_2d_4x 🔒

minecraft:overworld/caves/spaghetti_2d (SIMD form, batches 4 Y values)

compute_overworld__caves__spaghetti_2d_thickness_modulator 🔒

minecraft:overworld/caves/spaghetti_2d_thickness_modulator

compute_overworld__caves__spaghetti_2d_thickness_modulator_4x 🔒

minecraft:overworld/caves/spaghetti_2d_thickness_modulator (SIMD form, batches 4 Y values)

compute_overworld__caves__spaghetti_roughness_function 🔒

minecraft:overworld/caves/spaghetti_roughness_function

compute_overworld__caves__spaghetti_roughness_function_4x 🔒

minecraft:overworld/caves/spaghetti_roughness_function (SIMD form, batches 4 Y values)

compute_overworld__continents 🔒

minecraft:overworld/continents

compute_overworld__depth 🔒

minecraft:overworld/depth

compute_overworld__depth_4x 🔒

minecraft:overworld/depth (SIMD form, batches 4 Y values)

compute_overworld__erosion 🔒

minecraft:overworld/erosion

compute_overworld__factor 🔒

minecraft:overworld/factor

compute_overworld__jaggedness 🔒

minecraft:overworld/jaggedness

compute_overworld__offset 🔒

minecraft:overworld/offset

compute_overworld__ridges 🔒

minecraft:overworld/ridges

compute_overworld__ridges_folded 🔒

minecraft:overworld/ridges_folded

compute_overworld__sloped_cheese 🔒

minecraft:overworld/sloped_cheese

compute_overworld__sloped_cheese_4x 🔒

minecraft:overworld/sloped_cheese (SIMD form, batches 4 Y values)

compute_router_continentalness 🔒

compute_router_erosion 🔒

compute_router_preliminary_surface_level 🔒

compute_router_ridges 🔒

compute_router_temperature 🔒

compute_router_vegetation 🔒

compute_shift_x 🔒

minecraft:shift_x

compute_shift_z 🔒

minecraft:shift_z

compute_y 🔒

minecraft:y

compute_y_4x 🔒

minecraft:y (SIMD form, batches 4 Y values)

fill_cell_corner_densities

Evaluate the inner functions of all Interpolated markers at a cell corner.

fill_cell_corner_densities_4x

SIMD form of fill_cell_corner_densities that batches 4 cell-corner Y values at fixed (x, z).

router_barrier

Noise router entry: barrier

router_continentalness

Noise router entry: continentalness

router_depth

Noise router entry: depth

router_erosion

Noise router entry: erosion

router_final_density

Noise router entry: final_density

router_fluid_level_floodedness

Noise router entry: fluid_level_floodedness

router_fluid_level_spread

Noise router entry: fluid_level_spread

router_lava

Noise router entry: lava

router_preliminary_surface_level

Noise router entry: preliminary_surface_level

router_ridges

Noise router entry: ridges

router_temperature

Noise router entry: temperature

router_vegetation

Noise router entry: vegetation

router_vein_gap

Noise router entry: vein_gap

router_vein_ridged

Noise router entry: vein_ridged

router_vein_toggle

Noise router entry: vein_toggle

spline_helper_0 🔒

spline_helper_1 🔒

spline_helper_2 🔒

spline_helper_3 🔒

spline_helper_4 🔒

spline_helper_5 🔒

spline_helper_6 🔒

spline_helper_7 🔒

spline_helper_8 🔒

spline_helper_9 🔒

spline_helper_10 🔒

spline_helper_11 🔒

spline_helper_12 🔒

spline_helper_13 🔒

spline_helper_14 🔒

spline_helper_15 🔒

spline_helper_16 🔒

spline_helper_17 🔒

spline_helper_18 🔒

spline_helper_19 🔒

spline_helper_20 🔒

spline_helper_21 🔒

spline_helper_22 🔒

spline_helper_23 🔒

spline_helper_24 🔒

spline_helper_25 🔒

spline_helper_26 🔒

spline_helper_27 🔒

spline_helper_28 🔒

spline_helper_29 🔒

spline_helper_30 🔒

spline_helper_31 🔒

spline_helper_32 🔒

spline_helper_33 🔒

spline_helper_34 🔒

spline_helper_35 🔒

spline_helper_36 🔒

spline_helper_37 🔒

spline_helper_38 🔒

spline_helper_39 🔒

spline_helper_40 🔒

spline_helper_41 🔒

spline_helper_42 🔒

spline_helper_43 🔒

spline_helper_44 🔒

spline_helper_45 🔒

spline_helper_46 🔒

spline_helper_47 🔒

spline_helper_48 🔒

spline_helper_49 🔒

spline_helper_50 🔒

spline_helper_51 🔒

spline_helper_52 🔒

spline_helper_53 🔒

spline_helper_54 🔒

spline_helper_55 🔒

spline_helper_56 🔒

spline_helper_57 🔒

spline_helper_58 🔒

spline_helper_59 🔒

spline_helper_60 🔒

spline_helper_61 🔒

spline_helper_62 🔒

spline_helper_63 🔒

spline_helper_64 🔒

spline_helper_65 🔒

spline_helper_66 🔒

spline_helper_67 🔒

spline_helper_68 🔒

spline_helper_69 🔒

spline_helper_70 🔒

spline_helper_71 🔒

spline_helper_72 🔒

spline_helper_73 🔒

spline_helper_74 🔒

spline_helper_75 🔒

spline_helper_76 🔒

spline_helper_77 🔒

spline_helper_78 🔒

spline_helper_79 🔒

spline_helper_80 🔒

spline_helper_81 🔒

spline_helper_82 🔒

spline_helper_83 🔒

spline_helper_84 🔒

spline_helper_85 🔒

spline_helper_86 🔒

spline_helper_87 🔒

spline_helper_88 🔒

spline_helper_89 🔒

spline_helper_90 🔒

spline_helper_91 🔒

spline_helper_92 🔒

spline_helper_93 🔒

spline_helper_94 🔒

spline_helper_95 🔒

spline_helper_96 🔒

spline_helper_97 🔒

spline_helper_98 🔒

spline_helper_99 🔒

spline_helper_100 🔒

spline_helper_101 🔒

spline_helper_102 🔒

spline_helper_103 🔒

spline_helper_104 🔒

spline_helper_105 🔒

spline_helper_106 🔒

spline_helper_107 🔒

spline_helper_108 🔒

spline_helper_109 🔒

spline_helper_110 🔒

spline_helper_111 🔒

spline_helper_112 🔒

spline_helper_113 🔒

spline_helper_114 🔒