steel_worldgen::density_functions::overworld

Struct OverworldNoises

pub struct OverworldNoises {

    pub n_aquifer_barrier: NormalNoise,
    pub n_aquifer_fluid_level_floodedness: NormalNoise,
    pub n_aquifer_fluid_level_spread: NormalNoise,
    pub n_aquifer_lava: NormalNoise,
    pub n_cave_cheese: NormalNoise,
    pub n_cave_entrance: NormalNoise,
    pub n_cave_layer: NormalNoise,
    pub n_continentalness: NormalNoise,
    pub n_erosion: NormalNoise,
    pub n_jagged: NormalNoise,
    pub n_noodle: NormalNoise,
    pub n_noodle_ridge_a: NormalNoise,
    pub n_noodle_ridge_b: NormalNoise,
    pub n_noodle_thickness: NormalNoise,
    pub n_offset: NormalNoise,
    pub n_ore_gap: NormalNoise,
    pub n_ore_vein_a: NormalNoise,
    pub n_ore_vein_b: NormalNoise,
    pub n_ore_veininess: NormalNoise,
    pub n_pillar: NormalNoise,
    pub n_pillar_rareness: NormalNoise,
    pub n_pillar_thickness: NormalNoise,
    pub n_ridge: NormalNoise,
    pub n_spaghetti_2d: NormalNoise,
    pub n_spaghetti_2d_elevation: NormalNoise,
    pub n_spaghetti_2d_modulator: NormalNoise,
    pub n_spaghetti_2d_thickness: NormalNoise,
    pub n_spaghetti_3d_1: NormalNoise,
    pub n_spaghetti_3d_2: NormalNoise,
    pub n_spaghetti_3d_rarity: NormalNoise,
    pub n_spaghetti_3d_thickness: NormalNoise,
    pub n_spaghetti_roughness: NormalNoise,
    pub n_spaghetti_roughness_modulator: NormalNoise,
    pub n_temperature: NormalNoise,
    pub n_vegetation: NormalNoise,
    pub blended_noise: BlendedNoise,

}

All noise generators needed by this dimension’s density functions.

Created at runtime from a seed via the create method.

Fields
n_aquifer_barrier: NormalNoisen_aquifer_fluid_level_floodedness: NormalNoisen_aquifer_fluid_level_spread: NormalNoisen_aquifer_lava: NormalNoisen_cave_cheese: NormalNoisen_cave_entrance: NormalNoisen_cave_layer: NormalNoisen_continentalness: NormalNoisen_erosion: NormalNoisen_jagged: NormalNoisen_noodle: NormalNoisen_noodle_ridge_a: NormalNoisen_noodle_ridge_b: NormalNoisen_noodle_thickness: NormalNoisen_offset: NormalNoisen_ore_gap: NormalNoisen_ore_vein_a: NormalNoisen_ore_vein_b: NormalNoisen_ore_veininess: NormalNoisen_pillar: NormalNoisen_pillar_rareness: NormalNoisen_pillar_thickness: NormalNoisen_ridge: NormalNoisen_spaghetti_2d: NormalNoisen_spaghetti_2d_elevation: NormalNoisen_spaghetti_2d_modulator: NormalNoisen_spaghetti_2d_thickness: NormalNoisen_spaghetti_3d_1: NormalNoisen_spaghetti_3d_2: NormalNoisen_spaghetti_3d_rarity: NormalNoisen_spaghetti_3d_thickness: NormalNoisen_spaghetti_roughness: NormalNoisen_spaghetti_roughness_modulator: NormalNoisen_temperature: NormalNoisen_vegetation: NormalNoiseblended_noise: BlendedNoise
Implementations

impl OverworldNoises

pub fn create( seed: u64, splitter: &RandomSplitter, params: &FxHashMap<String, NoiseParameters>, ) -> Self

Create all noise generators from a world seed, positional splitter, and noise parameters.

impl OverworldNoises

fn apply_surface_rule_impl( ctx: &mut SurfaceRuleContext<'_>, ) -> Option<BlockStateId>

Apply this dimension’s surface rule at the current context position.